requirements.txt
dnspython~=2.2.1
idna~=3.3
pip==22.0.4
wheel~=0.37.1
pydantic~=1.9.0
Django~=4.0.4
sqlparse~=0.4.2
asgiref~=3.5.0
setuptools==62.1.0
fhir.resources